+++ NEWS.md
@@ -62,6 +62,7 @@ remove deprecated functions. The highlights below are grouped by theme.
   
 ### New data sources and collections
 * Add AlphaEarth (AEF) as a data source
+* Add support for data cubes shared on HuggingFace (`source = "HF:<user>"`)
 * Add support for the Amazonia-1 and GLAD image collections in the Brazil Data
   Cube (BDC)
 * Add support for the `SENTINEL-1-GRD` collection in CDSE